+++ ui/components/shadcn/index.ts
@@ -18,4 +18,5 @@ export * from "./separator/separator";
 export * from "./skeleton/skeleton";
 export * from "./tabs/generic-tabs";
 export * from "./tabs/tabs";
+export * from "./textarea/textarea";
 export * from "./tooltip";